refactor(security): share pinned artifact policy
